fix: moved js files to script directory
