+++ src/index.js
@@ -1,3 +1,3 @@
 
 import "./pages/index.css";
-import "./script";
+import "./scripts/script";
