+++ src/scripts/script.js
