+++ webpack.config.js
@@ -24,14 +24,6 @@ module.exports = {
                     'css-loader', 'postcss-loader']
             },
             {
-                test: /\.(png|jpe?g|gif)$/i,
-                use: [
-                    {
-                        loader: 'file-loader',
-                    },
-                ],
-            },
-            {
                 test: /\.(png|jpg|gif|ico|svg)$/i,
                 use: [
                     {
